Updated build script and added People

+++ clean.js
@@ -4,6 +4,14 @@ const path = require('path')
 const directory = 'dist'
 
 console.log(`Cleaning ${directory}...`)
+const WHITELIST = [
+	"pwabuilder-sw.js",
+	"manifest.json",
+	"timing.requestframe.worker.js",
+	"timing.setinterval.worker.js",
+	"timing.settimeout.worker.js"
+
+]
 
 fs.readdir(directory, (err, files) => {
   if (err) throw err
@@ -16,9 +24,18 @@ fs.readdir(directory, (err, files) => {
 	  {
 
 	  }else{
-		fs.unlink(location, err => {
-			if (err) throw err
-		})
+
+		// check to see if it is whitelisted
+		if (WHITELIST.indexOf(file) > -1)
+		{
+			// ignore
+		}else{
+			// DELETE
+			fs.unlink(location, err => {
+				if (err) throw err
+			})	
+		}
+	
 	  }
    
   }
